URL-based share button
This commit updates the 'Share' button (which currently depends
on a discontinued JSON-storage backend service) to save the
editor state in a base64-encoded URL query parameter.

For Firefox, Chrome, and Safari, this should pose no problem
(those browsers support URLs tens of thousands of characters
long). MS Edge limits URLs to 2000 characters. The editor currently
warns the user if their share link is that long (this corresponds to
a medium-to-large sized grammar).